Added edit comment functionality

+++ handlers/deletecomment.py
@@ -23,6 +23,14 @@ class DeleteComment(Handler):
     def post(self, comment_id):
         # get comment and delete it
         postcomment = Comments.get_by_id(int(comment_id))
-        logging.info(postcomment)
-        postcomment.delete()
+
+        #confirm existance of the comment
+        if not postcomment:
+            self.error(404)
+            return
+        user = self.user.name
+        post_author = postcomment.user
+        #confirm the post exists and the user is the author
+        if postcomment and user == post_author:
+            postcomment.delete()
