+++ handlers/deletepost.py
@@ -20,6 +20,15 @@ class DeletePost(Handler):
             return
 
     def post(self, post_id):
-        #get post and delete it
+        """delete the post"""
+        #check user is logged in
+        if not self.user:
+            self.redirect("/login")
+            return
+
         post = Post.get_by_id(int(post_id), parent=blog_key())
-        post.delete()
+        user = self.user.name
+        post_author = post.author
+        #confirm the post exists and the user is the author
+        if post and user == post_author:
+            post.delete()
